+++ tests/sdk/v2/api/test_wf_run.py
@@ -16,6 +16,8 @@ from unittest.mock import DEFAULT, MagicMock, Mock, PropertyMock, create_autospe
 import pytest
 
 from orquestra.sdk._base import _api, _workflow, serde
+from orquestra.sdk._base._spaces._api import list_projects, list_workspaces
+from orquestra.sdk._base._spaces._structs import ProjectRef, Workspace
 from orquestra.sdk._base.abc import RuntimeInterface
 from orquestra.sdk.exceptions import (
     ProjectInvalidError,
@@ -28,7 +30,7 @@ from orquestra.sdk.exceptions import (
 from orquestra.sdk.schema import ir
 from orquestra.sdk.schema.configs import RuntimeName
 from orquestra.sdk.schema.local_database import StoredWorkflowRun
-from orquestra.sdk.schema.workflow_run import ProjectRef, RunStatus, State
+from orquestra.sdk.schema.workflow_run import RunStatus, State
 from orquestra.sdk.schema.workflow_run import TaskRun as TaskRunModel
 
 from ..data.complex_serialization.workflow_defs import (
@@ -792,3 +794,76 @@ class TestProjectId:
                 "in_process", workspace_id=workspace_id, project_id=project_id
             )
             assert wf._project == expected
+
+
+class TestListWorkspaces:
+    @staticmethod
+    @pytest.fixture
+    def mock_config_runtime(monkeypatch):
+        ws = MagicMock()
+        type(ws).workspace_id = PropertyMock(
+            side_effect=[
+                "ws1",
+                "ws2",
+            ]
+        )
+        runtime = Mock(RuntimeInterface)
+        # For getting workflow ID
+        runtime.list_workspaces.return_value = [ws, ws]
+        mock_config = MagicMock(_api.RuntimeConfig)
+        mock_config._get_runtime.return_value = runtime
+        monkeypatch.setattr(
+            _api.RuntimeConfig, "load", MagicMock(return_value=mock_config)
+        )
+
+        return runtime
+
+    def test_list_workspaces(self, mock_config_runtime):
+        # Given
+        # When
+        runs = list_workspaces("mocked_config")
+        # Then
+        assert len(runs) == 2
+        assert runs[0].workspace_id == "ws1"
+        assert runs[1].workspace_id == "ws2"
+        mock_config_runtime.list_workspaces.assert_called_once()
+
+
+class TestListProjects:
+    @staticmethod
+    @pytest.fixture
+    def mock_config_runtime(monkeypatch):
+        ws = MagicMock()
+        type(ws).project_id = PropertyMock(
+            side_effect=[
+                "p1",
+                "p2",
+            ]
+        )
+        runtime = Mock(RuntimeInterface)
+        # For getting workflow ID
+        runtime.list_projects.return_value = [ws, ws]
+        mock_config = MagicMock(_api.RuntimeConfig)
+        mock_config._get_runtime.return_value = runtime
+        monkeypatch.setattr(
+            _api.RuntimeConfig, "load", MagicMock(return_value=mock_config)
+        )
+
+        return runtime
+
+    @pytest.mark.parametrize(
+        "workspace, expected_argument",
+        [
+            ("string_workspace_id", "string_workspace_id"),
+            (Workspace(workspace_id="id", name="name"), "id"),
+        ],
+    )
+    def test_list_projects(self, mock_config_runtime, workspace, expected_argument):
+        # Given
+        # When
+        runs = list_projects("mocked_config", workspace)
+        # Then
+        assert len(runs) == 2
+        assert runs[0].project_id == "p1"
+        assert runs[1].project_id == "p2"
+        mock_config_runtime.list_projects.assert_called_with(expected_argument)